cleaning up code layout
