+++ components/share-results.tsx
@@ -1,0 +1,35 @@
+import { cn } from "@/lib/utils";
+import { buttonVariants } from "./ui/button";
+import { Popover, PopoverContent, PopoverTrigger } from "./ui/popover";
+
+
+interface ShareResultsProps {
+    date: Date;
+    reveals: { value: number; used: boolean }[];
+    isCorrect: boolean | null;
+}
+
+const ShareResults = ({date, reveals, isCorrect}: ShareResultsProps) => {
+
+    return (
+
+        <Popover>
+        <PopoverTrigger
+            className={cn(
+                buttonVariants({ variant: "outline" }),
+                "w-5/6 bg-black text-white"
+            )}
+            onClick={() => {
+                const revealEmojis = "🔍".repeat(reveals.filter(reveal => reveal.used).length);
+                const resultEmoji = isCorrect ? "🦁" : "🤷";
+                navigator.clipboard.writeText(`${date.toDateString()}\nLiger Results\n${revealEmojis}${resultEmoji}`);
+            }}
+        >
+            Share Results
+        </PopoverTrigger>
+        <PopoverContent>Results copied to clipboard!</PopoverContent>
+    </Popover>
+    )
+}
+
+export default ShareResults;
